+++ app/components/user/row.js
@@ -1,8 +1,8 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
-import { inject as service } from '@ember/service';
+import { service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
-import { string, test } from 'yup';
+import { string } from 'yup';
 
 export default class UserRowComponent extends Component {
   @service constants;