+++ ember-intl.config.mjs
@@ -1,0 +1,8 @@
+export default {
+  buildOptions: {
+    fallbackLocale: undefined,
+    inputPath: 'translations',
+    publicOnly: false,
+    wrapTranslationsWithNamespace: true,
+  },
+}